requirements.txt
absl-py==1.4.0
adal==1.2.7
alembic==1.12.0
applicationinsights==0.11.10
argcomplete==2.1.2
astunparse==1.6.3
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
category-encoders==2.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.4
distro==1.8.0
docker==6.1.3
dotnetcore2==3.1.23
Flask==2.3.3
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
gast==0.4.0
gunicorn==21.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.14
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.12.0
knack==0.10.1
libclang==16.0.6
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.0
pandas==1.5.3
paramiko==3.3.1
pathspec==0.11.2
patsy==0.5.3
pkginfo==1.9.6
portalocker==2.8.2
protobuf==4.24.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rembg==2.0.50
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.2
SecretStorage==3.3.3
six==1.16.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
statsmodels==0.14.0
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.4
werkzeug==2.3.7
wrapt==1.14.1
zipp==3.16.2
psycopg2-binary